move remaining plotting functions into utils.plot
