+++ bundlewrap/utils/plot.py
@@ -1,3 +1,6 @@
+import re
+
+from . import names
 from .text import mark_for_translation as _, red
 
 
@@ -103,6 +106,91 @@ def graph_for_items(
     yield "fontsize = 28"
     yield "label = \"{}\"".format(title)
     yield "labelloc = \"t\""
+    yield "}"
+
+
+def plot_group(groups, nodes, show_nodes):
+    yield "digraph bundlewrap"
+    yield "{"
+
+    # Print subgraphs *below* each other
+    yield "rankdir = LR"
+
+    # Global attributes
+    yield ("node [color=\"#303030\"; "
+                 "fillcolor=\"#303030\"; "
+                 "fontname=Helvetica]")
+    yield "edge [arrowhead=vee]"
+
+    for group in groups:
+        yield "\"{}\" [fontcolor=white,style=filled];".format(group.name)
+
+    for node in nodes:
+        yield "\"{}\" [fontcolor=\"#303030\",shape=box,style=rounded];".format(node.name)
+
+    for group in groups:
+        for subgroup in group.immediate_subgroup_names:
+            yield "\"{}\" -> \"{}\" [color=\"#6BB753\",penwidth=2]".format(group.name, subgroup)
+        for subgroup in group._subgroup_names_from_patterns:
+            yield "\"{}\" -> \"{}\" [color=\"#6BB753\",penwidth=2]".format(group.name, subgroup)
+
+    if show_nodes:
+        for group in groups:
+            for node in group._nodes_from_members:
+                yield "\"{}\" -> \"{}\" [color=\"#D18C57\",penwidth=2]".format(
+                    group.name, node.name)
+
+            for node in group._nodes_from_patterns:
+                yield "\"{}\" -> \"{}\" [color=\"#714D99\",penwidth=2]".format(
+                    group.name, node.name)
+
+            for node in nodes:
+                if group in node._groups_dynamic:
+                    yield "\"{}\" -> \"{}\" [color=\"#FF0000\",penwidth=2]".format(
+                        group.name, node.name)
+
+    yield "}"
+
+
+def plot_node_groups(node):
+    yield "digraph bundlewrap"
+    yield "{"
+
+    # Print subgraphs *below* each other
+    yield "rankdir = LR"
+
+    # Global attributes
+    yield ("node [color=\"#303030\"; "
+                 "fillcolor=\"#303030\"; "
+                 "fontname=Helvetica]")
+    yield "edge [arrowhead=vee]"
+
+    for group in node.groups:
+        yield "\"{}\" [fontcolor=white,style=filled];".format(group.name)
+
+    yield "\"{}\" [fontcolor=\"#303030\",shape=box,style=rounded];".format(node.name)
+
+    for group in node.groups:
+        for subgroup in group.immediate_subgroup_names:
+            if subgroup in names(node.groups):
+                yield "\"{}\" -> \"{}\" [color=\"#6BB753\",penwidth=2]".format(group.name, subgroup)
+        for pattern in group.immediate_subgroup_patterns:
+            compiled_pattern = re.compile(pattern)
+            for group2 in node.groups:
+                if compiled_pattern.search(group2.name) is not None and group2 != group:
+                    yield "\"{}\" -> \"{}\" [color=\"#6BB753\",penwidth=2]".format(group.name, group2.name)
+
+    for group in node.groups:
+        if node in group._nodes_from_members:
+            yield "\"{}\" -> \"{}\" [color=\"#D18C57\",penwidth=2]".format(
+                group.name, node.name)
+        elif node in group._nodes_from_patterns:
+            yield "\"{}\" -> \"{}\" [color=\"#714D99\",penwidth=2]".format(
+                group.name, node.name)
+        elif group in node._groups_dynamic:
+            yield "\"{}\" -> \"{}\" [color=\"#FF0000\",penwidth=2]".format(
+                group.name, node.name)
+
     yield "}"
